+++ Tests/EditorLogicTests/EditorLogicTests.cpp
@@ -169,6 +169,8 @@ int main(int argc, char** argv)
 		testReferenceCardDerivesSharedPresentationState();
 		testVSTBusModelMigratesAndEdits();
 		testVSTSlotFillModel();
+		testChannelFlow();
+		testVSTRowDocument();
 		testPanelMonitorGateOpensOnlyForSelfGeneratedAudio();
 		testPanelMonitorGateArmDelayOutlivesReverbTails();
 		testPanelMonitorGateClosesAndRearms();
@@ -179,6 +181,7 @@ int main(int argc, char** argv)
 		testFilterCommandCatalogIconsExistOnDisk();
 		testFilterCommandCatalogTemplateRoster();
 		testFilterCommandCatalogDescriptions();
+		testFilterCommandCatalogChannelSelectionTargets();
 		testFilterPickerModelMatchesTermsAndPreservesCatalogIndices();
 		testFilterPickerModelOwnsSelectionNavigation();
 		testSharedRawBodyAndRoutingViewPredicates();